+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fix(fetchMap): getColorAccessor treats colorMap as override of actual domain (#238)
+
 ## 0.5
 
 ### 0.5.17